move fees to common mod
