+++ committer/src/main.rs
@@ -7,7 +7,7 @@ mod setup;
 use api::launch_api_server;
 use errors::{Result, WithContext};
 use metrics::prometheus::Registry;
-use services::fee_metrics_tracker::cache::CachingApi;
+use services::fees::cache::CachingApi;
 use setup::last_finalization_metric;
 use tokio_util::sync::CancellationToken;
