+++ fee_algo_simulation/src/state.rs
@@ -1,4 +1,4 @@
-use services::fee_metrics_tracker::cache::CachingApi;
+use services::fees::cache::CachingApi;
 
 /// Shared state across routes.
 #[derive(Clone)]
