+++ packages/services/src/fee_metrics_tracker.rs
@@ -1,5 +1,1 @@
-pub mod cache;
-pub mod port;
 pub mod service;
-#[cfg(feature = "test-helpers")]
-pub mod testing;
