+++ packages/test-helpers/src/lib.rs
@@ -11,9 +11,9 @@ use rand::{Rng, RngCore};
 use services::{
     block_committer::service::BlockCommitter,
     block_importer::service::BlockImporter,
-    fee_metrics_tracker::{
-        port::l1::Fees,
+    fees::{
         testing::{ConstantFeeApi, PreconfiguredFeeApi},
+        Fees,
     },
     state_listener::service::StateListener,
     types::{BlockSubmission, CollectNonEmpty, CompressedFuelBlock, Fragment, L1Tx, NonEmpty},
